+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.33] 2022-03-21
+## [0.5.33] 2022-03-23
 
 ### Added
 
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - (Fixed issue #74) Initialization of animation tracks was fixed.
+
+### Changed
+
+- Now **CAnimationCurve**'s track uses tangents instead of control points to change its curvature.
 
 ## [0.5.32] 2022-03-16